Lexical (static) scope.
I realize now why dynamic scope in a functional language is a **bad**
idea. When passing a function into another function, e.g. 'if', the
implementation of 'if' can affect how the passed function works, since
the local variables in the higher-order function can shadow variables
that the passed-in function expects to see.

So Airball is now lexically scoped.